requirements.txt
alembic==1.9.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.0
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
